+++ src/lib.rs
@@ -358,17 +358,10 @@ pub mod prelude {
     pub use bevy_enhanced_input_macros::{InputAction, InputContext};
 }
 
-use bevy::{
-    ecs::{
-        system::{ParamBuilder, QueryParamBuilder},
-        world::FilteredEntityMut,
-    },
-    input::InputSystem,
-    prelude::*,
-};
+use bevy::{input::InputSystem, prelude::*};
 
-use action_instances::{ActionInstances, ActionsRegistry};
-use input_reader::{ActionSources, InputReader, ResetInput};
+use action_instances::ContextRegistry;
+use input_reader::{ActionSources, ResetInput};
 use prelude::*;
 
 /// Initializes contexts and feeds inputs to them.
@@ -378,8 +371,7 @@ pub struct EnhancedInputPlugin;
 
 impl Plugin for EnhancedInputPlugin {
     fn build(&self, app: &mut App) {
-        app.init_resource::<ActionInstances>()
-            .init_resource::<ActionsRegistry>()
+        app.init_resource::<ContextRegistry>()
             .init_resource::<ResetInput>()
             .init_resource::<ActionSources>()
             .configure_sets(PreUpdate, EnhancedInputSystem.after(InputSystem));
@@ -388,70 +380,17 @@ impl Plugin for EnhancedInputPlugin {
     fn finish(&self, app: &mut App) {
         let registry = app
             .world_mut()
-            .remove_resource::<ActionsRegistry>()
+            .remove_resource::<ContextRegistry>()
             .expect("registry should be inserted in `build`");
 
-        let update = (
-            ParamBuilder,
-            ParamBuilder,
-            ParamBuilder,
-            ParamBuilder,
-            QueryParamBuilder::new(|builder| {
-                builder.optional(|builder| {
-                    for &id in registry.iter() {
-                        builder.mut_id(id);
-                    }
-                });
-            }),
-        )
-            .build_state(app.world_mut())
-            .build_system(update);
-
-        let rebuild = (
-            ParamBuilder,
-            ParamBuilder,
-            ParamBuilder,
-            ParamBuilder,
-            QueryParamBuilder::new(|builder| {
-                builder.optional(|builder| {
-                    for &id in registry.iter() {
-                        builder.mut_id(id);
-                    }
-                });
-            }),
-        )
-            .build_state(app.world_mut())
-            .build_any_system(rebuild);
-
-        app.add_observer(rebuild)
-            .add_systems(PreUpdate, update.in_set(EnhancedInputSystem));
+        for contexts in registry.iter() {
+            contexts.setup(app);
+        }
     }
-}
-
-fn update(
-    mut commands: Commands,
-    mut reader: InputReader,
-    time: Res<Time<Virtual>>, // We explicitly use `Virtual` to have access to `relative_speed`.
-    mut instances: ResMut<ActionInstances>,
-    mut actions: Query<FilteredEntityMut>,
-) {
-    reader.update_state();
-    instances.update(&mut commands, &mut reader, &time, &mut actions);
-}
-
-fn rebuild(
-    _trigger: Trigger<RebuildBindings>,
-    mut commands: Commands,
-    mut reset_input: ResMut<ResetInput>,
-    mut instances: ResMut<ActionInstances>,
-    time: Res<Time<Virtual>>,
-    mut actions: Query<FilteredEntityMut>,
-) {
-    instances.rebuild(&mut commands, &mut reset_input, &time, &mut actions);
 }
 
 /// Label for the system that updates input context instances.
 ///
-/// Runs in [`PreUpdate`].
+/// Runs in each registered [`InputContext::Schedule`].
 #[derive(Debug, PartialEq, Eq, Clone, Hash, SystemSet)]
 pub struct EnhancedInputSystem;