+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `InputContext::Schedule` to control the schedule in which the context will be evaluated.
+
 ## [0.9.0] - 2025-04-08
 
 ### Added